fix an issue in the usage of 2 error codes  + switch the location and change the names of 2 error classes  + move errors from web3-eth to web3-errors an update the imports referring them
