+++ packages/web3-eth/src/utils/get_transaction_gas_pricing.ts
@@ -19,7 +19,7 @@ import { FormatType, DataFormat, format, ETH_DATA_FORMAT } from 'web3-utils';
 import { Web3Context } from 'web3-core';
 import { EthExecutionAPI, Numbers, Transaction } from 'web3-types';
 import { isNullish } from 'web3-validator';
-import { Eip1559NotSupportedError, UnsupportedTransactionTypeError } from '../errors';
+import { Eip1559NotSupportedError, UnsupportedTransactionTypeError } from 'web3-errors';
 // eslint-disable-next-line import/no-cycle
 import { getBlock, getGasPrice } from '../rpc_method_wrappers';
 import { InternalTransaction } from '../types';
