+++ packages/web3-eth/src/utils/transaction_builder.ts
@@ -36,13 +36,13 @@ import { Web3Context } from 'web3-core';
 import { privateKeyToAddress } from 'web3-eth-accounts';
 import { getId, Web3NetAPI } from 'web3-net';
 import { isNullish, isNumber } from 'web3-validator';
-import { NUMBER_DATA_FORMAT } from '../constants';
 import {
 	InvalidTransactionWithSender,
 	LocalWalletNotAvailableError,
 	TransactionDataAndInputError,
 	UnableToPopulateNonceError,
-} from '../errors';
+} from 'web3-errors';
+import { NUMBER_DATA_FORMAT } from '../constants';
 // eslint-disable-next-line import/no-cycle
 import { getChainId, getTransactionCount } from '../rpc_method_wrappers';
 import { detectTransactionType } from './detect_transaction_type';
