+++ packages/web3-eth/test/fixtures/validation.ts
@@ -25,7 +25,7 @@ import {
 	TransactionCall,
 	TransactionWithSenderAPI,
 } from 'web3-types';
-import { InvalidTransactionCall, InvalidTransactionWithSender } from '../../src/errors';
+import { InvalidTransactionCall, InvalidTransactionWithSender } from 'web3-errors';
 
 export const isBaseTransactionValidData: [BaseTransactionAPI, true][] = [
 	[
