+++ packages/web3-eth/test/unit/validate_transaction_for_signing.test.ts
@@ -17,7 +17,7 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 
 /* eslint-disable jest/no-conditional-expect */
 
-import { InvalidTransactionObjectError } from '../../src/errors';
+import { InvalidTransactionObjectError } from 'web3-errors';
 import { validateTransactionForSigning } from '../../src/validation';
 import {
 	invalidNonceOrChainIdData,
